Add Artsy Remote Docker Circle CI Orb
This orb establishes a connection with an Artsy-managed remote Docker
daemon in order to run a Docker build in an environment optimized to
retain cached Docker layers.

If the orb is unable to establish a connection with the Artsy-managed
remote Docker daemon, the Docker build will fallback to Circle CI's
infrastructure.