Update validate_orb to take either path to orb or orb name
Previously validate_orb.sh took only the orb name. This was to make it simple to run. So it might be invoked like `./scripts/validate_orb.sh yarn` to validate the `yarn` orb. Given that I'm looking to add commit hook validation, I need it to also be able to support paths. Now it checks to see if the ORB input is a path to an orb file. If so, it sets that to `ORB_PATH` and uses a new `get_orb_from_path` method to reformat the path into the `ORB` name value. This is the method that will be used by `lint-staged`.