chore: bump remote docker version to 20.10.16
This PR bumps the remote docker version for hokusai to 20.10.16.
The reason why I'd like to update the version is that the Ruby alpine image above version 3.14 can't be built with docker below version 20.10.0. I'm trying to update the ruby version in Exchange to 3.1.2 and the ruby 3.1.2 alpine image is provided with above version 3.14. So in order to update the ruby version in Exchange to 3.1.2, only my option is to update the docker version. But it affects all of our services that use the hokusai image, so I decided not to update the docker version to the latest one.